Module symbol_map::indexing

Indexing on top of a Table.

It is anticipated that most uses cases will be covered by HashIndexing, which owns a Table and provides bidirectional mappings between data values and their symbols.

The Indexing trait is provided in case another lookup method is needed.

Structs

HashIndexing
HashMap-backed table indexing.

Ref
Wrapper for a raw pointer which lets us treat it like a reference.

Enums

Insertion
Indicates whether the result of a symbol lookup had to create a new table entry.

Traits

Indexing
Provides indexing for a Table, so that its elements may be retrieved efficiently. Most table lookups should go through an implementation of this trait structure instead of a Table directly.